+++ phoenix-core/src/test/java/org/apache/phoenix/transaction/TransactionServiceManager.java
@@ -1,0 +1,38 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.phoenix.transaction;
+
+import java.sql.SQLException;
+
+import org.apache.hadoop.conf.Configuration;
+import org.apache.phoenix.jdbc.PhoenixEmbeddedDriver.ConnectionInfo;
+import org.apache.phoenix.transaction.TransactionFactory.Provider;
+
+public class TransactionServiceManager {
+
+    public static PhoenixTransactionService startTransactionService(TransactionFactory.Provider provider, Configuration config, ConnectionInfo connInfo, int port) throws SQLException {
+        PhoenixTransactionProvider transactionProvider = provider.getTransactionProvider();
+        if(provider == Provider.TEPHRA) {
+            return TephraTransactionService.startAndInjectTephraTransactionService((TephraTransactionProvider)transactionProvider, config, connInfo, port);
+        } else if (provider == Provider.OMID) {
+            return OmidTransactionService.startAndInjectOmidTransactionService((OmidTransactionProvider)transactionProvider, config, connInfo, port);
+        }
+        throw new UnsupportedOperationException("Unknown transaction provider");
+    }
+
+}
